fix: export podiumSend extension of Response
Also add some online documentation to the method, in addition to
"testing our types" using <reference> in the unit tests
